added KVO for UNUserNotifications delegate
