+++ Mixpanel/AutomaticEvents.swift
@@ -9,6 +9,7 @@
 
 protocol AEDelegate {
     func track(event: String?, properties: Properties?)
+    func trackPushNotification(_ userInfo: [AnyHashable: Any], event: String)
 }
 
 #if DECIDE
@@ -80,6 +81,7 @@ class AutomaticEvents: NSObject, SKPaymentTransactionObserver, SKProductsRequest
                                                object: nil)
 
         SKPaymentQueue.default().add(self)
+        setupAutomaticPushTracking()
 
     }
 
@@ -154,6 +156,82 @@ class AutomaticEvents: NSObject, SKPaymentTransactionObserver, SKProductsRequest
         objc_sync_exit(awaitingTransactions)
     }
 
+    func setupAutomaticPushTracking() {
+        guard let appDelegate = MixpanelInstance.sharedUIApplication()?.delegate else {
+            return
+        }
+        var selector: Selector? = nil
+        var newSelector: Selector? = nil
+        let aClass: AnyClass = type(of: appDelegate)
+        var newClass: AnyClass?
+        if #available(iOS 10.0, *), let UNDelegate = UNUserNotificationCenter.current().delegate {
+            newClass = type(of: UNDelegate)
+        } else if #available(iOS 10.0, *) {
+            UNUserNotificationCenter.current().addDelegateObserver(ae: self)
+        }
+
+        if let newClass = newClass,
+            #available(iOS 10.0, *),
+            class_getInstanceMethod(newClass,
+                                    NSSelectorFromString("userNotificationCenter:willPresentNotification:withCompletionHandler:")) != nil {
+            selector = NSSelectorFromString("userNotificationCenter:willPresentNotification:withCompletionHandler:")
+            newSelector = #selector(UIResponder.UserNotificationCenter(_:newWillPresent:withCompletionHandler:))
+        } else if class_getInstanceMethod(aClass, NSSelectorFromString("application:didReceiveRemoteNotification:fetchCompletionHandler:")) != nil {
+            selector = NSSelectorFromString("application:didReceiveRemoteNotification:fetchCompletionHandler:")
+            newSelector = #selector(UIResponder.application(_:newDidReceiveRemoteNotification:fetchCompletionHandler:))
+        } else if class_getInstanceMethod(aClass, NSSelectorFromString("application:didReceiveRemoteNotification:")) != nil {
+            selector = NSSelectorFromString("application:didReceiveRemoteNotification:")
+            newSelector = #selector(UIResponder.application(_:newDidReceiveRemoteNotification:))
+        }
+
+        if let selector = selector, let newSelector = newSelector {
+            let block = { (view: AnyObject?, command: Selector, param1: AnyObject?, param2: AnyObject?) in
+                if let param2 = param2 as? [AnyHashable: Any] {
+                    self.delegate?.trackPushNotification(param2, event: "$campaign_received")
+                }
+            }
+            Swizzler.swizzleSelector(selector,
+                                     withSelector: newSelector,
+                                     for: aClass,
+                                     name: "notification opened",
+                                     block: block)
+        }
+    }
+
+    override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
+        if #available(iOS 10.0, *),
+            keyPath == "delegate",
+            let UNDelegate = UNUserNotificationCenter.current().delegate {
+            let delegateClass: AnyClass = type(of: UNDelegate)
+            if class_getInstanceMethod(delegateClass,
+                                       NSSelectorFromString("userNotificationCenter:willPresentNotification:withCompletionHandler:")) != nil {
+                let selector = NSSelectorFromString("userNotificationCenter:willPresentNotification:withCompletionHandler:")
+                let newSelector = #selector(UIResponder.UserNotificationCenter(_:newWillPresent:withCompletionHandler:))
+                let block = { (view: AnyObject?, command: Selector, param1: AnyObject?, param2: AnyObject?) in
+                    if let param2 = param2 as? [AnyHashable: Any] {
+                        self.delegate?.trackPushNotification(param2, event: "$campaign_received")
+                    }
+                }
+                Swizzler.swizzleSelector(selector,
+                                         withSelector: newSelector,
+                                         for: delegateClass,
+                                         name: "notification opened",
+                                         block: block)
+            }
+        }
+    }
+
+    deinit {
+        removeObserver(self, forKeyPath: "UNUserNotificationCenter.current().delegate")
+    }
+
+}
+
+@available(iOS 10.0, *)
+extension UNUserNotificationCenter {
+    func addDelegateObserver(ae: AutomaticEvents) {
+        addObserver(ae, forKeyPath: "delegate", options: [.old, .new], context: nil)
+    }
 }
 
 extension UIResponder {
