+++ MixpanelDemo/MixpanelDemo/AppDelegate.swift
@@ -18,6 +18,14 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey : Any]? = nil) -> Bool {
         var ADD_YOUR_MIXPANEL_TOKEN_BELOW_🛠🛠🛠🛠🛠🛠: String
+        Mixpanel.initialize(token: "MIXPANEL_TOKEN")
+        Mixpanel.mainInstance().loggingEnabled = true
+        Mixpanel.mainInstance().flushInterval = 5
+        let allTweaks: [TweakClusterType] = [MixpanelTweaks.floatTweak,
+                                             MixpanelTweaks.intTweak,
+                                             MixpanelTweaks.boolTweak,
+                                             MixpanelTweaks.stringTweak]
+        MixpanelTweaks.setTweaks(tweaks: allTweaks)
 
         if #available(iOS 10.0, *) {
             UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound, .badge], completionHandler: { (granted, error) in
@@ -31,15 +39,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
             UIApplication.shared.registerUserNotificationSettings(settings)
             UIApplication.shared.registerForRemoteNotifications()
         }
-
-        Mixpanel.initialize(token: "MIXPANEL_TOKEN")
-        Mixpanel.mainInstance().loggingEnabled = true
-        Mixpanel.mainInstance().flushInterval = 5
-        let allTweaks: [TweakClusterType] = [MixpanelTweaks.floatTweak,
-                                             MixpanelTweaks.intTweak,
-                                             MixpanelTweaks.boolTweak,
-                                             MixpanelTweaks.stringTweak]
-        MixpanelTweaks.setTweaks(tweaks: allTweaks)
 
         Mixpanel.mainInstance().identify(
             distinctId: Mixpanel.mainInstance().distinctId)
